+++ app/services/rails_xapi/statement_creator.rb
@@ -10,6 +10,7 @@ class RailsXapi::StatementCreator < ApplicationService
 
   def call(async: false)
     statement = prepare_statement
+    # Use a job for asynchronous calls.
     return RailsXapi::CreateStatementJob.perform_now(statement) if async
 
     statement.save
@@ -19,7 +20,7 @@ class RailsXapi::StatementCreator < ApplicationService
   private
 
   def prepare_statement
-    actor = RailsXapi::Actor.build_actor_from_data(@actor || @data[:actor])
+    actor = RailsXapi::Actor.build_actor_from_data(@actor.presence || @data[:actor])
 
     verb = RailsXapi::Verb.find_or_create_by(id: @data[:verb][:id]) do |v|
       v.display = @data[:verb][:display]
@@ -38,7 +39,7 @@ class RailsXapi::StatementCreator < ApplicationService
       object: object,
       result: result,
       context: context,
-      timestamp: @data[:timestamp]
+      timestamp: @data[:timestamp].presence || Time.zone.now
     )
     raise RailsXapi::Errors::XapiError, statement.errors.full_messages.join(", ") unless statement.valid?