+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## 0.1.2
+
+### Breaking Changes:
+
+- Removed `call_async` method from `RailsXapi::StatementCreator`, in favor of merging it into the `call` method.